Rework current-meal's probationary effect specs (#296) (#302)
